Added transitions for the header
Header will now be fixed at the top on scrolling and will shrinking slightly as scrolled down and back to normal size once scrolled closer back to the top. Removed some margins and padding for the nav and h1 and included line-height instead as it's a slightly simpler change when the transition happens.