Not quite ready yet
Still the issue of scotchPanel attributes interfering with the non-scotchPanel layout especially after window resizing. Needs to only be functional when in mobile view, outside of that it has no purpose. On ready, the inital styles will be correct, but after the going past the width breakpoint, the opposites attr will be removed and the opposite view width will be broken - mainly on scroll down for desktop and for mobile the panel will pop out abruptly.