A sort of fix to the panel and scroll issue
just added !important - it works - to have the transform be none on > 48em view. The js functions I had made would remove the attr, yet when they were re-added on window resize, they'd be missing attributes except the ones that broke down the scroll and panel functionality. There is a better way