Edited the pwsp height attr
Changed to 100vh as for whatever reason, pwsp was reading the height as the entire length of the page so once clicked on an image that was say, first, that image would be displayed in the middle of the whole scale of the page rather than centered wherever the viewer was at. It's not perfect as once clicked, the image is centered on your view, while the background jumps back to the top of the page. After closing, the background - which is now your view - goes back to where you had clicked. Works great on desktop. Internet browsing on your phone is shit let's be honest.